+++ app/scraper/outlets/horas24.py
@@ -15,6 +15,7 @@ from bs4 import BeautifulSoup
 
 from scraper.base import HARD_PAGE_CAP
 from scraper.base_playwright import BasePlaywrightScraper
+from scraper.utils import bare_term
 
 logger = logging.getLogger(__name__)
 
@@ -56,7 +57,7 @@ class Horas24Scraper(BasePlaywrightScraper):
         urls: list[str] = []
         for page in range(1, HARD_PAGE_CAP + 1):
             params = dict(_SEARCH_DEFAULTS)
-            params["search_texto"] = phrase
+            params["search_texto"] = bare_term(phrase)
             params["search_pag"] = page
             try:
                 resp = _requests.get(